Rename from u_arizona to uarizona to match what integration initialization script would produce
