+++ src/reformatters/contrib/uarizona/swann/analysis/templates/latest.zarr/snow_depth/zarr.json
