+++ src/reformatters/contrib/uarizona/swann/analysis/templates/latest.zarr/snow_water_equivalent/zarr.json
