+++ src/reformatters/contrib/uarizona/swann/analysis/templates/latest.zarr/spatial_ref/zarr.json
