+++ src/reformatters/contrib/uarizona/swann/analysis/templates/latest.zarr/time/zarr.json
